+++ src/pages/SearchByName.jsx
@@ -42,6 +42,8 @@ const SearchByName = () => {
     }
 
     const handleDelete = async (id) => {
+        const confirmation = window.confirm('Are you sure you want to delete this consumer?')
+        if(!confirmation) return
         try{
             await axios.delete(process.env.REACT_APP_BACKEND_URL+'consumer/'+id, {
                 headers:{
